Renamed controls to control
