+++ frictionless/__init__.py
@@ -1,5 +1,6 @@
 from .check import Check
 from .config import VERSION as __version__
+from .control import Control
 from .describe import *
 from .extract import *
 from .field import Field
@@ -28,7 +29,6 @@ from .transform import *
 from .type import Type
 from .validate import *
 from . import checks
-from . import controls
 from . import dialects
 from . import errors
 from . import exceptions
